ci: Replace tag workflow with a shell script
Realistically this would be a pain to implement since actions from
workflows (like pushing a git tag) cannot trigger other workflows (like
building a container image).

In theory it would be possible to do it all in a single workflow with a
lot of conditions, but I might as well just tag the damn thing manually.

Signed-off-by: Manuel Hutter <manuel@hutter.io>